refactor: merge host, hosts and host-manager packages into single package
