+++ apps/host/package.json
@@ -14,7 +14,7 @@
   },
   "dependencies": {
     "@yuants/protocol": "workspace:*",
-    "@yuants/host-manager": "workspace:*",
+    "@yuants/utils": "workspace:*",
     "rxjs": "~7.5.6",
     "ws": "~8.8.1"
   },
