--- apps/hosts/CHANGELOG.md
@@ -1,360 +1,0 @@
-# Change Log - @yuants/app-hosts
-
-This log was last generated on Sat, 23 Aug 2025 20:08:09 GMT and should not be manually modified.
-
-## 0.1.63
-Sat, 23 Aug 2025 20:08:09 GMT
-
-_Version update only_
-
-## 0.1.62
-Tue, 05 Aug 2025 16:36:43 GMT
-
-_Version update only_
-
-## 0.1.61
-Fri, 01 Aug 2025 13:20:25 GMT
-
-_Version update only_
-
-## 0.1.60
-Sat, 12 Jul 2025 09:58:56 GMT
-
-_Version update only_
-
-## 0.1.59
-Sat, 12 Jul 2025 06:20:10 GMT
-
-_Version update only_
-
-## 0.1.58
-Fri, 11 Jul 2025 19:02:39 GMT
-
-_Version update only_
-
-## 0.1.57
-Thu, 10 Jul 2025 01:26:47 GMT
-
-_Version update only_
-
-## 0.1.56
-Tue, 24 Jun 2025 15:34:54 GMT
-
-_Version update only_
-
-## 0.1.55
-Tue, 24 Jun 2025 12:18:24 GMT
-
-_Version update only_
-
-## 0.1.54
-Sun, 15 Jun 2025 06:01:16 GMT
-
-_Version update only_
-
-## 0.1.53
-Tue, 10 Jun 2025 20:44:01 GMT
-
-_Version update only_
-
-## 0.1.52
-Mon, 26 May 2025 07:50:33 GMT
-
-_Version update only_
-
-## 0.1.51
-Sun, 27 Apr 2025 15:56:48 GMT
-
-_Version update only_
-
-## 0.1.50
-Mon, 21 Apr 2025 17:48:12 GMT
-
-### Patches
-
-- upgrade node version to 22
-
-## 0.1.49
-Fri, 18 Apr 2025 18:31:55 GMT
-
-### Patches
-
-- add Dockerfile node option
-
-## 0.1.48
-Mon, 07 Apr 2025 16:39:44 GMT
-
-_Version update only_
-
-## 0.1.47
-Tue, 01 Apr 2025 22:55:16 GMT
-
-_Version update only_
-
-## 0.1.46
-Thu, 27 Mar 2025 08:28:07 GMT
-
-_Version update only_
-
-## 0.1.45
-Thu, 13 Mar 2025 01:35:22 GMT
-
-_Version update only_
-
-## 0.1.44
-Tue, 11 Mar 2025 15:25:28 GMT
-
-_Version update only_
-
-## 0.1.43
-Mon, 10 Mar 2025 20:30:46 GMT
-
-_Version update only_
-
-## 0.1.42
-Sat, 08 Mar 2025 21:00:28 GMT
-
-_Version update only_
-
-## 0.1.41
-Wed, 19 Feb 2025 18:08:11 GMT
-
-_Version update only_
-
-## 0.1.40
-Tue, 21 Jan 2025 19:55:44 GMT
-
-_Version update only_
-
-## 0.1.39
-Sun, 19 Jan 2025 20:56:17 GMT
-
-_Version update only_
-
-## 0.1.38
-Sat, 18 Jan 2025 11:58:42 GMT
-
-_Version update only_
-
-## 0.1.37
-Tue, 14 Jan 2025 12:10:20 GMT
-
-_Version update only_
-
-## 0.1.36
-Mon, 13 Jan 2025 17:14:27 GMT
-
-_Version update only_
-
-## 0.1.35
-Mon, 13 Jan 2025 15:36:18 GMT
-
-_Version update only_
-
-## 0.1.34
-Mon, 13 Jan 2025 14:31:04 GMT
-
-_Version update only_
-
-## 0.1.33
-Wed, 08 Jan 2025 19:37:36 GMT
-
-_Version update only_
-
-## 0.1.32
-Thu, 02 Jan 2025 07:08:15 GMT
-
-### Patches
-
-- upgrade container base image to bookworm for wrtc
-
-## 0.1.31
-Thu, 19 Dec 2024 15:28:50 GMT
-
-_Version update only_
-
-## 0.1.30
-Tue, 17 Dec 2024 08:15:54 GMT
-
-_Version update only_
-
-## 0.1.29
-Sat, 07 Dec 2024 14:42:13 GMT
-
-### Patches
-
-- dependency error
-
-## 0.1.28
-Sat, 07 Dec 2024 02:19:50 GMT
-
-### Patches
-
-- extract lib @yuants/host-manager
-
-## 0.1.27
-Wed, 04 Dec 2024 09:36:47 GMT
-
-_Version update only_
-
-## 0.1.26
-Wed, 27 Nov 2024 15:48:19 GMT
-
-_Version update only_
-
-## 0.1.25
-Sun, 24 Nov 2024 06:27:35 GMT
-
-_Version update only_
-
-## 0.1.24
-Sun, 24 Nov 2024 05:13:22 GMT
-
-_Version update only_
-
-## 0.1.23
-Sat, 23 Nov 2024 15:55:38 GMT
-
-_Version update only_
-
-## 0.1.22
-Sat, 23 Nov 2024 05:37:50 GMT
-
-### Patches
-
-- change default resource settings
-
-## 0.1.21
-Tue, 19 Nov 2024 10:43:22 GMT
-
-### Patches
-
-- add hash bucket 
-
-## 0.1.20
-Sat, 16 Nov 2024 06:09:31 GMT
-
-_Version update only_
-
-## 0.1.19
-Wed, 13 Nov 2024 12:49:50 GMT
-
-### Patches
-
-- chore: add deploy priority and affinity
-
-## 0.1.18
-Sat, 02 Nov 2024 10:25:33 GMT
-
-_Version update only_
-
-## 0.1.17
-Sat, 26 Oct 2024 08:08:45 GMT
-
-### Patches
-
-- fix missing terminal
-
-## 0.1.16
-Fri, 25 Oct 2024 21:53:01 GMT
-
-### Patches
-
-- fix phantom terminal
-
-## 0.1.15
-Thu, 24 Oct 2024 19:57:43 GMT
-
-_Version update only_
-
-## 0.1.14
-Sun, 29 Sep 2024 12:31:38 GMT
-
-### Patches
-
-- fix phantom terminal
-
-## 0.1.13
-Wed, 25 Sep 2024 08:29:52 GMT
-
-_Version update only_
-
-## 0.1.12
-Tue, 17 Sep 2024 20:07:02 GMT
-
-_Version update only_
-
-## 0.1.11
-Tue, 17 Sep 2024 14:54:15 GMT
-
-_Version update only_
-
-## 0.1.10
-Sun, 15 Sep 2024 14:00:45 GMT
-
-_Version update only_
-
-## 0.1.9
-Sun, 15 Sep 2024 05:36:59 GMT
-
-_Version update only_
-
-## 0.1.8
-Mon, 09 Sep 2024 14:05:59 GMT
-
-_Version update only_
-
-## 0.1.7
-Sun, 08 Sep 2024 10:16:35 GMT
-
-_Version update only_
-
-## 0.1.6
-Sat, 07 Sep 2024 20:34:59 GMT
-
-### Patches
-
-- delete normal closed terminal
-
-## 0.1.5
-Sat, 07 Sep 2024 19:24:07 GMT
-
-### Patches
-
-- RetryError Maximum call stack size exceeded
-
-## 0.1.4
-Sat, 07 Sep 2024 18:05:13 GMT
-
-### Patches
-
-- Phantom Terminal Elimination
-
-## 0.1.3
-Wed, 04 Sep 2024 12:44:10 GMT
-
-### Patches
-
-- update private key name
-
-## 0.1.2
-Mon, 02 Sep 2024 18:18:05 GMT
-
-### Patches
-
-- add list hosts as an admin method
-
-## 0.1.1
-Fri, 30 Aug 2024 16:26:37 GMT
-
-### Patches
-
-- fix ingress error
-
-## 0.1.0
-Fri, 30 Aug 2024 14:29:33 GMT
-
-### Minor changes
-
-- new
-
