--- apps/hosts/etc/app-hosts.api.md
@@ -1,9 +1,0 @@
-## API Report File for "@yuants/app-hosts"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-// (No @packageDocumentation comment for this package)
-
-```
