--- apps/hosts/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "@yuants/app-hosts",
-  "version": "0.1.63",
-  "description": "Hosts",
-  "files": [
-    "dist",
-    "lib",
-    "temp"
-  ],
-  "scripts": {
-    "build": "heft test --clean && api-extractor run --local && yuan-toolkit post-build"
-  },
-  "dependencies": {
-    "@yuants/protocol": "workspace:*",
-    "@yuants/utils": "workspace:*",
-    "@yuants/host-manager": "workspace:*",
-    "rxjs": "~7.5.6",
-    "ws": "~8.8.1"
-  },
-  "devDependencies": {
-    "@microsoft/api-extractor": "~7.30.0",
-    "@rushstack/heft": "~0.47.5",
-    "@rushstack/heft-jest-plugin": "~0.16.8",
-    "@rushstack/heft-node-rig": "~1.10.7",
-    "@types/heft-jest": "1.0.3",
-    "@types/json-schema": "~7.0.11",
-    "@types/node": "22",
-    "@types/ws": "~8.5.3",
-    "@yuants/extension": "workspace:*",
-    "@yuants/tool-kit": "workspace:*",
-    "typescript": "~4.7.4"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  }
-}
