--- apps/hosts/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./node_modules/@rushstack/heft-node-rig/profiles/default/tsconfig-base.json",
-  "compilerOptions": {
-    "types": ["heft-jest", "node"]
-  }
-}
