--- libraries/host-manager/CHANGELOG.md
@@ -1,202 +1,0 @@
-# Change Log - @yuants/host-manager
-
-This log was last generated on Sat, 23 Aug 2025 20:08:09 GMT and should not be manually modified.
-
-## 0.2.25
-Sat, 23 Aug 2025 20:08:09 GMT
-
-_Version update only_
-
-## 0.2.24
-Tue, 05 Aug 2025 16:36:43 GMT
-
-_Version update only_
-
-## 0.2.23
-Fri, 01 Aug 2025 13:20:25 GMT
-
-_Version update only_
-
-## 0.2.22
-Sat, 12 Jul 2025 09:58:56 GMT
-
-_Version update only_
-
-## 0.2.21
-Sat, 12 Jul 2025 06:20:10 GMT
-
-_Version update only_
-
-## 0.2.20
-Fri, 11 Jul 2025 19:02:39 GMT
-
-### Patches
-
-- move utils
-
-## 0.2.19
-Thu, 10 Jul 2025 01:26:47 GMT
-
-_Version update only_
-
-## 0.2.18
-Tue, 24 Jun 2025 15:34:54 GMT
-
-_Version update only_
-
-## 0.2.17
-Tue, 24 Jun 2025 12:18:24 GMT
-
-_Version update only_
-
-## 0.2.16
-Sun, 15 Jun 2025 06:01:16 GMT
-
-_Version update only_
-
-## 0.2.15
-Tue, 10 Jun 2025 20:44:01 GMT
-
-_Version update only_
-
-## 0.2.14
-Mon, 26 May 2025 07:50:33 GMT
-
-### Patches
-
-- apply for new metrics
-
-## 0.2.13
-Sun, 27 Apr 2025 15:56:48 GMT
-
-_Version update only_
-
-## 0.2.12
-Mon, 21 Apr 2025 17:48:12 GMT
-
-### Patches
-
-- upgrade node version to 22
-
-## 0.2.11
-Fri, 18 Apr 2025 18:31:55 GMT
-
-_Version update only_
-
-## 0.2.10
-Mon, 07 Apr 2025 16:39:44 GMT
-
-### Patches
-
-- clear label for metrics
-- fix label
-
-## 0.2.9
-Tue, 01 Apr 2025 22:55:16 GMT
-
-_Version update only_
-
-## 0.2.8
-Thu, 27 Mar 2025 08:28:07 GMT
-
-_Version update only_
-
-## 0.2.7
-Thu, 13 Mar 2025 01:35:22 GMT
-
-_Version update only_
-
-## 0.2.6
-Tue, 11 Mar 2025 15:25:28 GMT
-
-### Patches
-
-- remove provideChannel
-
-## 0.2.5
-Mon, 10 Mar 2025 20:30:46 GMT
-
-_Version update only_
-
-## 0.2.4
-Sat, 08 Mar 2025 21:00:28 GMT
-
-### Patches
-
-- update protocol api
-
-## 0.2.3
-Wed, 19 Feb 2025 18:08:11 GMT
-
-_Version update only_
-
-## 0.2.2
-Tue, 21 Jan 2025 19:55:44 GMT
-
-_Version update only_
-
-## 0.2.1
-Sun, 19 Jan 2025 20:56:17 GMT
-
-_Version update only_
-
-## 0.2.0
-Sat, 18 Jan 2025 11:58:42 GMT
-
-### Minor changes
-
-- add channel 'TerminalInfoChangeEvent'
-
-### Patches
-
-- disable Terminate
-
-## 0.1.7
-Tue, 14 Jan 2025 12:10:21 GMT
-
-_Version update only_
-
-## 0.1.6
-Mon, 13 Jan 2025 17:14:27 GMT
-
-_Version update only_
-
-## 0.1.5
-Mon, 13 Jan 2025 15:36:19 GMT
-
-_Version update only_
-
-## 0.1.4
-Mon, 13 Jan 2025 14:31:04 GMT
-
-_Version update only_
-
-## 0.1.3
-Wed, 08 Jan 2025 19:37:36 GMT
-
-_Version update only_
-
-## 0.1.2
-Thu, 02 Jan 2025 07:08:15 GMT
-
-_Version update only_
-
-## 0.1.1
-Thu, 19 Dec 2024 15:28:50 GMT
-
-_Version update only_
-
-## 0.1.0
-Tue, 17 Dec 2024 08:15:54 GMT
-
-### Minor changes
-
-- add some metrics
-
-## 0.0.1
-Sat, 07 Dec 2024 02:19:50 GMT
-
-### Patches
-
-- init package
-
