--- libraries/host-manager/config/jest.config.json
@@ -1,3 +1,0 @@
-{
-  "extends": "@rushstack/heft-node-rig/profiles/default/config/jest.config.json"
-}
