--- libraries/host-manager/etc/host-manager.api.md
@@ -1,45 +1,0 @@
-## API Report File for "@yuants/host-manager"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { ITerminalInfo } from '@yuants/protocol';
-import { Terminal } from '@yuants/protocol';
-import { default as WebSocket_2 } from 'ws';
-
-// @public (undocumented)
-export const createNodeJSHostManager: (config: IHostManagerConfig) => IHostManger;
-
-// @public (undocumented)
-export interface IHost {
-    // (undocumented)
-    dispose: () => void;
-    // (undocumented)
-    host_id: string;
-    // (undocumented)
-    host_terminal: Terminal;
-    // (undocumented)
-    mapTerminalIdToHasHeader: Record<string, boolean>;
-    // (undocumented)
-    mapTerminalIdToSocket: Record<string, WebSocket_2.WebSocket>;
-    // (undocumented)
-    terminalInfos: Map<string, ITerminalInfo>;
-}
-
-// @public (undocumented)
-export interface IHostManagerConfig {
-    mapHostUrlToHostId: (host_url: string) => string;
-}
-
-// @public (undocumented)
-export interface IHostManger {
-    // (undocumented)
-    dispose: () => void;
-    // (undocumented)
-    hosts: Record<string, IHost | undefined>;
-}
-
-// (No @packageDocumentation comment for this package)
-
-```
